Update required and nullable properties (ADV-23613)
Motivation
---
We want to update the isRequired and isNullable settings for the Add and Update schema object properties so that they will be excluded from the json output for non-critical fields.

Modifications
---
Updated the isRequired and isNullable settings for the Add and Update schema object properties so that only the most critical properties are required to have values, and the others will be excluded from the json output if they are null or empty.